backend/requirements.txt
# Web framework
fastapi
uvicorn[standard]
python-multipart

# Config
pydantic-settings
python-dotenv

# RAG & LLM
langchain
langchain-openai
langchain-google-genai
langchain-community
openai

# Vector store (Phase 1)
chromadb
faiss-cpu

# NLP / NER
spacy>=3.8.0
# scispacy (Fails to build on Python 3.13 due to spacy 3.6 dependency)
# https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_ner_bc5cdr_md-0.5.3.tar.gz

# OCR
pytesseract
Pillow
opencv-python-headless
transformers
torch

# Embeddings
sentence-transformers

# Async tasks
celery[redis]
redis
flower

# Database
sqlalchemy
asyncpg

# TTS & PDF Export
gTTS
reportlab

# Testing
pytest
pytest-asyncio
httpx

# Utilities
python-jose[cryptography]
structlog
